Fixing visual studio warnings

+++ LittleEngine/LittleEngine/Components/Transform.cpp
@@ -8,7 +8,7 @@ Transform::Transform(unsigned int shaderProgram) {
 	UpdateModel();
 }
 void Transform::UpdateModel(){
-	float movementInX = sin(Engine->moduleTimeController->gameTimeClock.Read()/1000) / 5;
+	double movementInX = sin(Engine->moduleTimeController->gameTimeClock.Read()/1000) / 5;
 	model = glm::translate(model, glm::vec3(movementInX,0.0f,0.0f));
 	glUniformMatrix4fv(modelShaderOutput, 1, GL_FALSE, glm::value_ptr(model));
 }
